# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para la gestión de la clínica veterinaria: autenticación, usuarios, clientes, empleados, mascotas, citas, pagos y notificaciones.",
        "title": "API Clínica Veterinaria",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Equipo de Soporte API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v2",
    "paths": {
...
        "/me/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated notifications for the authenticated user. Requires any authenticated role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get my notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated notifications",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.NotificationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/me/notifications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single notification by ID for the authenticated user. Requires any authenticated role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get my notification by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.NotificationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/notifications": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually sends a notification through the given channel (email, sms). Requires employee, manager, or admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Send notification (staff)",
                "parameters": [
                    {
                        "description": "Notification to send",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.SendNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification sent",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/notifications/channel/{channel}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated notifications filtered by delivery channel (email, sms). Requires employee, manager, or admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notifications by channel (staff)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delivery channel (email, sms)",
                        "name": "channel",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated notifications",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.NotificationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Missing or invalid channel",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/notifications/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a summary of sent notifications, optionally filtered by sender. Requires employee, manager, or admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification summary (staff)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sender ID to filter by",
                        "name": "sender_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification summary",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.NotificationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/notifications/type/{type}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated notifications filtered by type. Requires employee, manager, or admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notifications by type (staff)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated notifications",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.NotificationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Missing or invalid type",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single notification by ID. Requires employee, manager, or admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification by ID (staff)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.NotificationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "clinic-vet-api_internal_shared_http.APIResponse": {
            "description": "Standardized API response structure.",
            "type": "object",
            "properties": {
                "data": {
                    "description": "The data payload of the response for successful requests."
                },
                "error": {
                    "description": "Details of the error if the request was not successful.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.ErrorInfo"
                        }
                    ]
                },
                "message": {
                    "description": "A message providing additional context about the response.",
                    "type": "string"
                },
                "meta": {
                    "description": "Additional metadata for the response, such as pagination info."
                },
                "request_id": {
                    "description": "A unique identifier for the request.",
                    "type": "string"
                },
                "success": {
                    "description": "Indicates whether the request was successful.",
                    "type": "boolean"
                },
                "timestamp": {
                    "description": "The timestamp when the response was generated.",
                    "type": "string"
                }
            }
        },
        "clinic-vet-api_internal_shared_http.ErrorInfo": {
            "description": "Detailed information about an error.",
            "type": "object",
            "properties": {
                "code": {
                    "description": "The error code.",
                    "type": "string"
                },
                "details": {
                    "description": "A map of specific error details, often used for validation errors.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "A descriptive error message.",
                    "type": "string"
                },
                "stack": {
                    "description": "The error stack trace (for development purposes).",
                    "type": "string"
                },
                "type": {
                    "description": "The type of the error.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token. Example: \"Bearer \u0026lt;access_token\u0026gt;\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}